requirements.txt
streamlit
pandas
pdfplumber
google-generativeai